% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_9.R
\name{movingaverage}
\alias{movingaverage}
\title{Moving Average}
\usage{
movingaverage(x = x, side = "both", numbertoside = NULL)
}
\arguments{
\item{x}{a vector for which you will be calculating the moving average.}

\item{side}{whether you want to average elements on the left, right, or both
sides; should accept the following values: "left","right","both".}

\item{numbertoside}{the number of elements to the side of your index element
that you want to include in the moving average.}
}
\value{
A vector of the moving average
\item{Input}{input vector}
\item{Output}{output vector}
}
\description{
We created a function that calculates a moving average of a
  specified window width across a vector.
}
\details{
To create this function, we only used methods that were
  introduced in either Lecture 1 or Lecture 2.
}
\examples{
inputdata<-c(0,2.457,3.878,3.663,1.90,-0.658,-2.943,-3.986,-3.349,-1.299,
1.299,3.349,3.986,2.943,0.658,-1.904,-3.663,-3.878,-2.457,0)
movingaverage(x=inputdata,numbertoside=3,side="left")
}
