% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_9.R
\name{SimTtestPower}
\alias{SimTtestPower}
\title{Two-Sample Power Calculation Based on Simulation}
\usage{
SimTtestPower(Var1mean = NULL, Var2mean = NULL, Var1sd = NULL,
  Var2sd = NULL, Var1samplesize = NULL, Var2samplesize = NULL,
  nsim = 100, alphalevel = 0.05)
}
\arguments{
\item{Var1mean}{Variable 1 mean}

\item{Var2mean}{Variable 2 mean}

\item{Var1sd}{Variable 1 standard deviation}

\item{Var2sd}{Variable 2 standard deviation}

\item{Var1samplesize}{Variable 1 sample size}

\item{Var2samplesize}{Variable 2 sample size}

\item{nsim}{Number of simulations}

\item{alphalevel}{alpha-level (default=0.05)}
}
\value{
Empirical power calculation
}
\description{
Power is the probability we reject the null hypothesis given it
  is false. Building on this definition, create a function that uses
  simulations to estimate power for a two-sample T-test.
}
\details{
First, you will need to simulate two normally distributed variables,
  each with a distinct sample size, mean, and standard deviation, and perform
  a T-test. For that single simulation, evaluate if we would reject the null
  hypothesis given a specific alpha-level. Now repeat this simulation many
  times. Power can then be estimated as the proportion of simulations for
  which we rejected the null hypothesis.
}
\examples{
SimTtestPower(Var1mean=20,Var2mean=22,Var1sd=4,Var2sd=6,
Var1samplesize=40,Var2samplesize=40,nsim=10000,alphalevel=0.05)
}
